Move environment helper functions into own module
